`ia` now looks for plugins/commands installed to entry_point=internetarchive.cli.plugins. internetarchive.iacli renamed to internetarchive.cli.
