+++ tests/cli/test_ia_list.py
