+++ tests/cli/test_ia_metadata.py
