+++ tests/cli/test_ia_upload.py
